refactor: websocket repository
